refactor(ios-settings): extract model management section
- move the active model card, model rows, actions, provider selection, and model status helpers into a dedicated SettingsTabView+Models extension
- keep SettingsTabView focused on the remaining settings sections by removing the heaviest model-specific view logic
- register the new extension file in the Xcode project sources and views group
- preserve the not installed size labels for Whisper (~190 MB) and Parakeet (~480 MB) in the extracted model section
